全面优化：安全加固、性能提升、架构现代化、测试覆盖
安全修复（9 项）:
- 浏览量无限刷 → IP 限流（30次/60s）
- 限流器竞态条件 → DashMap 原子操作
- 文件删除/上传路径遍历 → safe_media_path canonicalize 校验
- 用户名长度限制（2-64）、注册时序侧信道修复
- Auth token 升级为 256-bit Alphanumeric
- Web XSS 修复（errorMsg 转义）、Token URL 泄露防护

后端优化:
- N+1 查询消除（check_existing_files 批量查询）
- 批量删除（单事务 DELETE WHERE id = ANY）
- 播放历史分页（LIMIT 500）
- 游标分页回填缩略图（不再加载全表）
- 连接池可配置（DB_MAX_CONNECTIONS）
- tracing 双写（stdout + JSON 文件）
- X-Request-ID 请求追踪
- 健康检查含 DB 探活
- OpenAPI 全端点文档化
- 4 个性能索引（migration 011）
- 101 个测试（单元 87 + 集成 14）

Android 优化:
- Jetpack Compose 重写 + Kawaii 暗色主题
- Koin 依赖注入（6 个 ViewModel + Repository）
- Room 离线缓存（网络优先 + 缓存降级）
- Compose BOM 升级至 2025.04.01
- ViewModel 统一（消除 AndroidViewModel）
- 分页加载（Home size=20 + loadMore）
- EncryptedSharedPreferences 缓存修复
- observeForever 内存泄漏修复
- 硬编码颜色集中管理（Color.kt）
- Log 生产环境守卫（BuildConfig.DEBUG）
- 搜索结果缩略图显示
- 6 个死代码 Fragment 删除（~1176 行）
- 废弃依赖清理（8 个）
- ProGuard 规则完善
- 仪器化 UI 测试 + 单元测试

Web PWA 优化:
- PWA 支持（manifest.json + Service Worker）
- CSS 压缩（atmos.min.css）+ 关键 CSS 内联
- 图片懒加载（IntersectionObserver）+ WebP 支持
- MD5 → Web Crypto SHA-256
- 共享模块提取（cards.js、selection.js）
- 焦点管理（trapFocus）+ ARIA 标签
- Toast 通知替代 alert()
- 键盘快捷键（J/K/L/F/M/0-9）
- 结构化错误日志（localStorage + window.__atmosErrors）
- 响应式图片（srcset thumb+cover）
- WCAG AA 对比度修复

文档:
- 新建 CLAUDE.md（中文，完整架构说明）
- 更新 AGENTS.md（修正过时信息）

+++ app/src/main/java/com/lanvideo/player/data/model/AuthModels.kt
@@ -33,7 +33,7 @@ data class UserProfileResponse(
     val username: String,
     val isAdmin: Boolean,
     val createdAt: String,
-    val totalVideosWatched: Int,
+    val totalVideosWatched: Long,
     val totalWatchTimeMs: Long,
     val recentHistory: List<RecentWatchItem>
 )
